Prefactor: restructure project e2e tests
This package will be used as a template for the next commit. Restructure it a bit beforehand.
